Remove unnecessary classes & update tests
