--- src/main/java/io/split/android/client/localhost/LocalhostAttributesManager.java
@@ -1,85 +1,0 @@
-package io.split.android.client.localhost;
-
-import static com.google.common.base.Preconditions.checkNotNull;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-
-import java.util.Map;
-
-import io.split.android.client.attributes.AttributesManager;
-import io.split.android.client.storage.attributes.AttributesStorage;
-import io.split.android.client.validators.AttributesValidator;
-import io.split.android.client.validators.ValidationMessageLogger;
-
-public class LocalhostAttributesManager implements AttributesManager {
-
-    private final AttributesStorage mAttributesStorage;
-    private final AttributesValidator mAttributesValidator;
-    private final ValidationMessageLogger mValidationMessageLogger;
-
-    public LocalhostAttributesManager(@NonNull AttributesStorage attributesStorage,
-                                      @NonNull AttributesValidator attributesValidator,
-                                      @NonNull ValidationMessageLogger validationMessageLogger) {
-        mAttributesStorage = checkNotNull(attributesStorage);
-        mAttributesValidator = checkNotNull(attributesValidator);
-        mValidationMessageLogger = checkNotNull(validationMessageLogger);
-    }
-
-    @Override
-    public boolean setAttribute(String attributeName, Object value) {
-        if (!mAttributesValidator.isValid(value)) {
-            logValidationWarning(attributeName);
-            return false;
-        }
-
-        mAttributesStorage.set(attributeName, value);
-
-        return true;
-    }
-
-    @Nullable
-    @Override
-    public Object getAttribute(String attributeName) {
-        return mAttributesStorage.get(attributeName);
-    }
-
-    @Override
-    public boolean setAttributes(Map<String, Object> attributes) {
-        for (Map.Entry<String, Object> attribute : attributes.entrySet()) {
-            if (!mAttributesValidator.isValid(attribute.getValue())) {
-                logValidationWarning(attribute.getKey());
-                return false;
-            }
-        }
-
-        mAttributesStorage.set(attributes);
-
-        return true;
-    }
-
-    private void logValidationWarning(String key) {
-        String mValidationTag = "split attributes";
-        mValidationMessageLogger.w("You passed an invalid attribute value for " + key + ", acceptable types are String, double, float, long, int, boolean or Collections", mValidationTag);
-    }
-
-    @NonNull
-    @Override
-    public Map<String, Object> getAllAttributes() {
-        return mAttributesStorage.getAll();
-    }
-
-    @Override
-    public boolean removeAttribute(String attributeName) {
-        mAttributesStorage.remove(attributeName);
-
-        return true;
-    }
-
-    @Override
-    public boolean clearAttributes() {
-        mAttributesStorage.clear();
-
-        return true;
-    }
-}
